Add files via upload
The CC.jsp is the main front page that takes user input: from currency, to currency and the amount.
The convertlogic.js does the conversion. The API used for this is free when the code was developed. It might become paid/use license key later.
The result is then displayed on the resultDisplay.jsp page using beans.

This could work as a simple beans project if required.